requirements.txt
streamlit
ephem
pillow
google-generativeai